chore: migrate functions to lerna
