--- firebase.json
@@ -1,5 +1,0 @@
-{
-  "functions": {
-    "predeploy": "npm --prefix functions run build"
-  }
-}
