+++ packages/functions/firebase.json
@@ -1,0 +1,5 @@
+{
+  "functions": {
+    "predeploy": "npm --prefix \"$RESOURCE_DIR\" run build"
+  }
+}
