+++ packages/functions/functions/tsconfig.json
@@ -2,12 +2,12 @@
   "compilerOptions": {
     "module": "commonjs",
     "noImplicitReturns": true,
+    "noUnusedLocals": true,
     "outDir": "lib",
     "sourceMap": true,
-    "target": "es2017",
-    "allowSyntheticDefaultImports": true,
-    "lib": ["es2017"]
+    "strict": true,
+    "target": "es2017"
   },
   "compileOnSave": true,
-  "include": ["src/**/*", "*.d.ts"]
+  "include": ["src"]
 }
