+++ packages/functions/package.json
@@ -1,28 +1,14 @@
 {
   "name": "functions",
   "scripts": {
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
+    "start": "npm run shell --prefix functions",
+    "postinstall": "npm install --prefix functions"
   },
   "main": "lib/index.js",
   "engines": {
     "node": "8"
   },
-  "dependencies": {
-    "firebase-admin": "~6.4.0",
-    "firebase-functions": "^2.3.1",
-    "node-fetch": "^2.6.0",
-    "xml2js": "^0.4.23"
-  },
-  "devDependencies": {
-    "@types/node-fetch": "^2.5.5",
-    "@types/xml2js": "^0.4.5",
-    "firebase-tools": "^6.12.0",
-    "typescript": "^3.8.3"
-  },
+  "dependencies": {},
+  "devDependencies": {},
   "private": true
 }
